+++ db/models/check.js
@@ -12,7 +12,7 @@ const Check = sequelize.define("check", {
     allowNull: false,
   },
   checkAmount: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(18, 2),
     defaultValue: 0,
   },
